requirements.txt
Django==5.2.12
aiogram==3.22.0
Telethon==1.41.2
python-dotenv==1.1.1
requests==2.32.5
gunicorn==23.0.0